Fix idle detection issue when a user has been logged on for a month.

+++ Procurement/Utility/PoeTradeOnlineHelper.cs
@@ -109,7 +109,7 @@ namespace Procurement.Utility
             };
             GetLastInputInfo(ref inputInfo);
             // Allow for TickCount wrap-around.
-            return TimeSpan.FromMilliseconds(unchecked(Environment.TickCount - inputInfo.dwTime));
+            return TimeSpan.FromMilliseconds(unchecked(Environment.TickCount - (int)inputInfo.dwTime));
         }
 
         internal void Start()
